Commit pop-up delete contract and pets

+++ FrontEnd/ProjetoStarterFrontEnd/src/app/pet-list/pet-list.component.ts
@@ -5,6 +5,8 @@ import "ag-grid-enterprise";
 import { ActionButtonComponent } from '../action-button/action-button.component';
 import { ActionButtonBeneficiariesComponent } from '../action-button-beneficiaries/action-button-beneficiaries.component';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { ConfirmationDialogComponent, ConfirmDialogModel } from '../components/shared/confirmation-dialog/confirmation-dialog.component';
+import { MatDialog } from '@angular/material';
 
 @Component({
   selector: 'app-pet-list',
@@ -12,7 +14,7 @@ import { MatSnackBar } from '@angular/material/snack-bar';
   styleUrls: ['./pet-list.component.scss']
 })
 export class PetListComponent implements OnInit {
-
+  public result: any;
   detailCellRendererParams;
   gridApi;
   gridColumApi;
@@ -21,31 +23,50 @@ export class PetListComponent implements OnInit {
   gridOptions: GridOptions;
   load_failure: boolean;
 
-  constructor(private http: HttpClient, private _snackBar: MatSnackBar) { }
+  constructor(public dialog: MatDialog, private http: HttpClient, private _snackBar: MatSnackBar) { }
+
+  confirmDialog(): void {
+    const message = `Are you sure you want to do this?`;
+
+    const dialogData = new ConfirmDialogModel("Confirm Action", message);
+
+    const dialogRef = this.dialog.open(ConfirmationDialogComponent, {
+      maxWidth: "400px",
+      data: dialogData
+    });
+
+    dialogRef.afterClosed().subscribe(dialogResult => {
+      this.result = dialogResult;
+    });
+  }
 
   ngOnInit() {
     this.setup_gridData();
     this.setup_gridOptions();
     this.paginationPageSize = 50;
   }
-  
+
   private handle_editUser(data: any) {
     //this.contractform.patchValue(data);
-    }
-  
-    private handle_deleteUser(data: any) {
-      console.log(data);
-      const id = data.beneficiaryId;
-      console.log(id);
+  }
+
+  private handle_deleteUser(data: any) {
+    console.log(data);
+    const id = data.beneficiaryId;
+    console.log(id);
+    this.confirmDialog();
+
+    if (this.result == true) {
       this.http.delete(`https://beneficiarieswebapi.azurewebsites.net/api/Beneficiary/${id}`).subscribe(response => this.setup_gridData(), error => this.openSnackBar(error.message), () => this.openSnackBar("Beneficiário removido com sucesso"));
     }
-      
-    openSnackBar(message: string): void {
-      this._snackBar.open(message, '', {
-        duration: 5000,
-        
-      });
-    }
+  }
+
+  openSnackBar(message: string): void {
+    this._snackBar.open(message, '', {
+      duration: 5000,
+
+    });
+  }
 
   private setup_gridOptions() {
 
@@ -73,7 +94,7 @@ export class PetListComponent implements OnInit {
           filter: true,
           cellRenderer: (data) => {
             return data.value ? (new Date(data.value)).toLocaleDateString() : '';
-          }, 
+          },
           onCellValueChanged:
             this.onCellEdit.bind(this)
         },
@@ -123,23 +144,23 @@ export class PetListComponent implements OnInit {
     const { data } = event;
   }
 }
-function SpeciesFormmatter(params){
+function SpeciesFormmatter(params) {
   return speciesValue(params.value);
 }
-function speciesValue(number){
-  if(number == 0){
+function speciesValue(number) {
+  if (number == 0) {
     return "Canis Lupus Familiaris";
   }
-  if(number == 1){
+  if (number == 1) {
     return "Felis Catus"
   }
-  if(number == 2){
-return "Mesocricetus Auratus"
+  if (number == 2) {
+    return "Mesocricetus Auratus"
   }
-  if(number == 3){
+  if (number == 3) {
     return "Nymphicus Hollandicus"
   }
-  if(number == 4){
+  if (number == 4) {
     return "Ara Chloropterus"
   }
 }
